updated the protocol
It seems that if the checksum results in 0x7e, it needs to be escaped
by doubling it (as not to be confused with the start of the next message,
apparently). Otherwise some funny glitches occur.

(Quite shitty programming, honestly. A receiver's state machine shall not be
affected by a certain value in the data stream.)

Signed-off-by: alxwk <awjd15@gmail.com>